Railsのエンドポイントをapi/v1から始まるようディレクトリ構成とルートを修正
