+++ config/routes.rb
@@ -2,5 +2,9 @@ Rails.application.routes.draw do
 # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   get '/users/index'
 
-  resources :ticker
+  namespace 'api' do
+    namespace 'v1' do
+      resources :ticker
+    end
+  end
 end
